+++ src/game/entities/entity.ts
@@ -4,11 +4,11 @@
 
   Base class for static map objects.
 */
-import {Vector2} from "./data_structures/vector2";
-import {MapObject} from "./map_object";
+import {Vector2} from "../data_structures/vector2";
+import {MapObject} from "../map_object";
 
 export class Entity extends MapObject {
   constructor(x: number, y: number) {
-    this.position = new Vector2(x, y);
+    this.setPositionV2(new Vector2(x, y));
   }
 }
