+++ test/entity_tests.ts
@@ -4,8 +4,8 @@
 
   Tests src/game/entity.ts
 */
-import {Entity} from "../src/game/entity";
-import {Vector2} from "../src/game/data_structures/vector2";
+import {Entity} from "../src/game/entities/entity";
+import {Vector2} from "../data_structures/vector2";
 
 var assert = require("assert");
 
@@ -14,14 +14,15 @@ describe("entity.ts", function() {
     it("should require a position to be instantiated and remain static", function() {
       //Test chunk 1
       var e = new Entity(1, 1);
-      assert.equal(e.position.x, 1, "chunk 1: x");
-      assert.equal(e.position.y, 1, "chunk 1: y");
+      var pos = e.getPosition();
+      assert.equal(e.getPosition().x, 1, "chunk 1: x");
+      assert.equal(e.getPosition().y, 1, "chunk 1: y");
     })
   }),
   describe("hp", function() {
     it("should have hp that increases/decreases", function() {
       //Test chunk 1
-      var e = new Entity();
+      var e = new Entity(1, 1);
       e.setHp(100);
       assert.equal(e.getHp(), 100, "chunk 1: hp start");
       e.addHp(-1);
